+++ app/frontend/frontend/components/CelestialObjects/Panel/index.vue
@@ -44,7 +44,7 @@ import type {
 import { SearchResultTypeEnum } from "@/services/fyApi";
 
 type Props = {
-  item: SearchResult;
+  celestialObject: CelestialObject;
 };
 
 const props = defineProps<Props>();
@@ -52,7 +52,7 @@ const props = defineProps<Props>();
 const resultItem = computed(() => props.item.item);
 
 const location = computed(
-  () => resultItem.value as CelestialObject | Starsystem,
+  () => resultItem.value as CelestialObject | Starsystem
 );
 
 const { supported: webpSupported } = useWebpCheck();