+++ app/models/ahoy/visit.rb
@@ -21,8 +21,9 @@
 #
 # Indexes
 #
-#  index_ahoy_visits_on_user_id      (user_id)
-#  index_ahoy_visits_on_visit_token  (visit_token) UNIQUE
+#  index_ahoy_visits_on_user_id                       (user_id)
+#  index_ahoy_visits_on_visit_token                   (visit_token) UNIQUE
+#  index_ahoy_visits_on_visitor_token_and_started_at  (visitor_token,started_at)
 #
 module Ahoy
   class Visit < ApplicationRecord