+++ config/routes/admin_routes.rb
@@ -62,11 +62,7 @@ namespace :admin, path: (Rails.configuration.app.on_subdomain? ? "" : "admin"), 
   resources :model_modules, path: "model-modules", except: [:show]
   resources :model_module_packages, path: "model-module-packages", except: [:show]
   resources :model_upgrades, path: "model-upgrades", except: [:show]
-  resources :model_paints, path: "model-paints", except: [:show] do
-    collection do
-      put :import
-    end
-  end
+  resources :model_paints, path: "model-paints", except: [:show]
 
   resources :manufacturers, except: [:show]