+++ db/migrate/20231004131207_add_missing_ahoy_index_on_started_at.rb
@@ -1,0 +1,5 @@
+class AddMissingAhoyIndexOnStartedAt < ActiveRecord::Migration[7.0]
+  def change
+    add_index :ahoy_visits, [:visitor_token, :started_at]
+  end
+end